implement authorization for backend

+++ hungry-belly-backend/src/main/java/com/eddie/hungry_belly_backend/scheduler/service/FoodImageCleanupService.java
@@ -2,29 +2,36 @@ package com.eddie.hungry_belly_backend.scheduler.service;
 
 import com.eddie.hungry_belly_backend.common.util.storage.dto.StorageObject;
 import com.eddie.hungry_belly_backend.common.util.storage.service.StorageService;
+import com.eddie.hungry_belly_backend.food.repository.FoodImageRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
 import java.time.Instant;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
-public class ExportCleanupService extends AbstractCleanupService {
+@Slf4j
+public class FoodImageCleanupService extends AbstractCleanupService {
+    private final FoodImageRepository imageRepository;
     private final StorageService storageService;
-    private static final String PREFIX = "exports/";
+
+    private static final String PREFIX = "food-photos/";
 
     @Override
     public void cleanupOrphanFiles() {
         List<StorageObject> files = storageService.listFiles(PREFIX);
+        Set<String> dbPaths = new HashSet<>(imageRepository.findAllPaths());
+
         Instant threshold = Instant.now().minus(Duration.ofHours(24));
 
         for(StorageObject file: files) {
             String filePath = file.getName();
-            if(file.getUpdatedAt().isBefore(threshold)) {
+            if(!dbPaths.contains(filePath) && file.getUpdatedAt().isBefore(threshold)) {
                 try {
                     storageService.deleteFile(file.getName());
                     log.info("Deleted orphan file: {}", filePath);
@@ -32,6 +39,7 @@ public class ExportCleanupService extends AbstractCleanupService {
                     log.error("Failed to delete file: {}", filePath, e);
                 }
             }
+
         }
     }
 }
